# app/static/swagger.yaml
swagger: "2.0"
...
info:
  title: "Mechanic Database"
  description: "API for managing a mechanic shop system, including customers, service tickets, mechanics, and inventory."
  version: "1.0.0"

# API Host and Schemes
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# Security Definitions
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"

# API Endpoints
paths:
...
  /customers:
    get:
      tags:
        - "customers"
      summary: "List all customers"
      description: "Retrieve a list of all customers (requires customer authentication)."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Customers retrieved successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CreateCustomerResponse"
    post:
      tags:
        - "customers"
      summary: "Create a new customer"
      description: "Register a new customer in the mechanic shop system."
      parameters:
        - in: "body"
          name: "body"
          description: "Customer registration information"
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created successfully"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
  /customers/login:
    post:
      tags:
        - "customers"
      summary: "Customer login"
      description: "Validates customer credentials and returns an authentication token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
  /customers/search:
    get:
      tags:
        - "customers"
      summary: "Search customers"
      description: "Search for customers by name or email (requires customer authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "query"
          name: "name"
          type: "string"
          required: true
          description: "Search query for name"
      responses:
        200:
          description: "Search results"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CreateCustomerResponse"
  /customers/{customer_id}:
    get:
      tags:
        - "customers"
      summary: "Get a customer"
      description: "Retrieve details of a specific customer by ID (requires customer authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "customer_id"
          type: "integer"
          required: true
          description: "ID of the customer"
      responses:
        200:
          description: "Customer retrieved successfully"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
    put:
      tags:
        - "customers"
      summary: "Update a customer"
      description: "Update an existing customer's information (requires customer authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "customer_id"
          type: "integer"
          required: true
          description: "ID of the customer"
        - in: "body"
          name: "body"
          description: "Updated customer information"
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        200:
          description: "Customer updated successfully"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
    delete:
      tags:
        - "customers"
      summary: "Delete a customer"
      description: "Remove a customer from the system (requires customer authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "customer_id"
          type: "integer"
          required: true
          description: "ID of the customer"
      responses:
        200:
          description: "Customer deleted successfully"
          schema:
            $ref: "#/definitions/DeleteResponse"

  # Service Ticket Endpoints
...
  /inventory:
    get:
      tags:
        - "inventory"
      summary: "List all parts"
      description: "Retrieve a list of all parts in inventory."
      responses:
        200:
          description: "Parts retrieved successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PartResponse"
    post:
      tags:
        - "inventory"
      summary: "Create a new part"
      description: "Add a new part to the inventory."
      parameters:
        - in: "body"
          name: "body"
          description: "Part information"
          required: true
          schema:
            $ref: "#/definitions/CreatePartPayload"
      responses:
        201:
          description: "Part created successfully"
          schema:
            $ref: "#/definitions/PartResponse"
  /inventory/{part_id}:
    get:
      tags:
        - "inventory"
      summary: "Get a part"
      description: "Retrieve details of a specific part by ID."
      parameters:
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true
          description: "ID of the part"
      responses:
        200:
          description: "Part retrieved successfully"
          schema:
            $ref: "#/definitions/PartResponse"
    put:
      tags:
        - "inventory"
      summary: "Update a part"
      description: "Update an existing part in inventory."
      parameters:
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true
          description: "ID of the part"
        - in: "body"
          name: "body"
          description: "Updated part information"
          required: true
          schema:
            $ref: "#/definitions/CreatePartPayload"
      responses:
        200:
          description: "Part updated successfully"
          schema:
            $ref: "#/definitions/PartResponse"
    delete:
      tags:
        - "inventory"
      summary: "Delete a part"
      description: "Remove a part from inventory."
      parameters:
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true
          description: "ID of the part"
      responses:
        200:
          description: "Part deleted successfully"
          schema:
            $ref: "#/definitions/DeleteResponse"
  /inventory/{part_id}/remove_stock:
    post:
      tags:
        - "inventory"
      summary: "Remove stock from a part"
      description: "Reduce the quantity of a specific part in stock."
      parameters:
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true
          description: "ID of the part"
        - in: "body"
          name: "body"
          description: "Quantity to remove"
          required: true
          schema:
            $ref: "#/definitions/RemoveStockPayload"
      responses:
        200:
          description: "Stock removed successfully"
          schema:
            $ref: "#/definitions/StockUpdateResponse"
  /inventory/{part_id}/add-to-ticket/{ticket_id}:
    post:
      tags:
        - "inventory"
      summary: "Add part to service ticket"
      description: "Add a part to a specific service ticket."
      parameters:
        - in: "path"
          name: "part_id"
          type: "integer"
          required: true
          description: "ID of the part"
        - in: "path"
          name: "ticket_id"
          type: "integer"
          required: true
          description: "ID of the service ticket"
      responses:
        200:
          description: "Part added to ticket successfully"

  # Utility Endpoints
...
definitions:
...
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      message:
        type: "string"
      status:
        type: "string"
      token:
        type: "string"

  # Customer Schemas
  CreateCustomerPayload:
    type: "object"
    properties:
      email:
        type: "string"
      name:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
    required:
      - email
      - name
      - password
      - phone
  CreateCustomerResponse:
    type: "object"
    properties:
      email:
        type: "string"
      id:
        type: "integer"
      name:
        type: "string"
      phone:
        type: "string"

  # Service Ticket Schemas
...
  CreatePartPayload:
    type: "object"
    properties:
      name:
        type: "string"
      price:
        type: "number"
      quantity_in_stock:
        type: "integer"
      description:
        type: "string"

    required:
      - name
      - price
      - quantity_in_stock
      - description
  PartResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      price:
        type: "number"
      quantity_in_stock:
        type: "integer"
      description:
        type: "string"
  RemoveStockPayload:
    type: "object"
    properties:
      quantity:
        type: "integer"
    required:
      - quantity
  StockUpdateResponse:
    type: "object"
    properties:
      message:
        type: "string"
      new_quantity_in_stock:
        type: "integer"

  # Utility Schemas
  DeleteResponse:
    type: "object"
    properties:
      message:
        type: "string"
...